Move changeMultiselect() to test/util/trigger.js
